Custom welcome message for new users to be made another day
The amount of work that has gone into this very tiny feature
is too much, and I'm literally no closer to implementing it than
before I started. (Though I learned many ways NOT to do it).

It's not even like the message to new users should be much
different than the one for returning users